split cms-blocks component
