+++ react/src/webshop/components/elements/cms-blocks/CmsBlocks.tsx
@@ -1,9 +1,10 @@
 import React, { Fragment } from 'react';
 import ImplementationPage from '../../../props/models/ImplementationPage';
 import Markdown from '../markdown/Markdown';
-import useAssetUrl from '../../../hooks/useAssetUrl';
 import classNames from 'classnames';
 import Section from '../sections/Section';
+import CmsBlocksItem from './CmsBlocksItem';
+import CmsBlocksItemLink from './CmsBlocksItemLink';
 
 export default function CmsBlocks({
     page,
@@ -12,8 +13,6 @@ export default function CmsBlocks({
     page: ImplementationPage;
     largeMarkdown?: boolean;
 }) {
-    const assetUrl = useAssetUrl();
-
     if (!page.description_html && page.blocks.length === 0) {
         return null;
     }
@@ -39,92 +38,9 @@ export default function CmsBlocks({
                         {page.blocks?.map((block) => (
                             <Fragment key={block.id}>
                                 {block.button_enabled && page.blocks_per_row > 1 ? (
-                                    <a
-                                        target={block.button_target_blank ? '_blank' : '_self'}
-                                        rel={block.button_target_blank ? 'noreferrer' : ''}
-                                        href={block.button_link}
-                                        className="fund-item">
-                                        <div className="fund-media">
-                                            <img
-                                                src={
-                                                    block.media?.sizes?.public ||
-                                                    assetUrl('/assets/img/placeholders/product-small.png')
-                                                }
-                                                alt=""
-                                            />
-                                        </div>
-                                        <div className="fund-information">
-                                            {block.title && <h2 className="fund-title">{block.title}</h2>}
-                                            {block.label && (
-                                                <div className="fund-label label label-primary">{block.label}</div>
-                                            )}
-
-                                            <div className="fund-description">
-                                                <Markdown content={block.description_html} />
-                                            </div>
-
-                                            <div className="fund-button">
-                                                <span className="button button-primary fund-button-button">
-                                                    {block.button_text}
-                                                    <div
-                                                        className="mdi mdi-arrow-right icon-right"
-                                                        aria-hidden="true"
-                                                    />
-                                                </span>
-                                                <span className="button button-text button-text-primary fund-button-link">
-                                                    {block.button_text}
-                                                </span>
-                                            </div>
-                                        </div>
-                                    </a>
+                                    <CmsBlocksItemLink block={block} />
                                 ) : (
-                                    <div className="fund-item">
-                                        <div className="fund-media">
-                                            <img
-                                                src={
-                                                    block.media?.sizes?.public ||
-                                                    assetUrl('/assets/img/placeholders/product-small.png')
-                                                }
-                                                alt=""
-                                            />
-                                        </div>
-                                        <div className="fund-information">
-                                            {block.title && <h2 className="fund-title">{block.title}</h2>}
-                                            {block.label && (
-                                                <div className="fund-label label label-primary">{block.label}</div>
-                                            )}
-
-                                            <div className="fund-description">
-                                                <Markdown content={block.description_html} />
-                                            </div>
-
-                                            {block.button_enabled && (
-                                                <div className="fund-button">
-                                                    <a
-                                                        className="button button-primary fund-button-button"
-                                                        target={block.button_target_blank ? '_blank' : '_self'}
-                                                        rel={block.button_target_blank ? 'noreferrer' : ''}
-                                                        href={block.button_link}>
-                                                        {block.button_text}
-                                                        <div
-                                                            className="mdi mdi-arrow-right icon-right"
-                                                            aria-hidden="true"
-                                                        />
-                                                    </a>
-                                                    {page.blocks_per_row > 1 && (
-                                                        <a
-                                                            className="button button-text button-text-primary fund-button-link"
-                                                            target={block.button_target_blank ? '_blank' : '_self'}
-                                                            rel={block.button_target_blank ? 'noreferrer' : ''}
-                                                            href={block.button_link}
-                                                            aria-label={block.button_link_label}>
-                                                            {block.button_text}
-                                                        </a>
-                                                    )}
-                                                </div>
-                                            )}
-                                        </div>
-                                    </div>
+                                    <CmsBlocksItem block={block} blocksPerRow={page.blocks_per_row} />
                                 )}
                             </Fragment>
                         ))}
